generate openapi swagger.json files
