+++ etc/proto/Dockerfile
@@ -20,6 +20,7 @@ RUN go install -v google.golang.org/protobuf/cmd/protoc-gen-go@v1.30.0
 RUN go install -v google.golang.org/grpc/cmd/protoc-gen-go-grpc@v1.3.0
 
 RUN go install -v github.com/grpc-ecosystem/grpc-gateway/v2/protoc-gen-grpc-gateway@v2.18.0
+RUN go install -v github.com/grpc-ecosystem/grpc-gateway/v2/protoc-gen-openapiv2@v2.18.0
 
 RUN mkdir -p ${GOPATH}/src/github.com/pachyderm/pachyderm
 RUN date +%s >/last_run_time
